WIP on plist encoding of Int128
XML plists are easy; just write out the string. Binary plists turn out to be subtle, however. The “obvious” encoding (array of two Int64s plus a sign) is stupidly inefficient; it is actually better to just use a hex string! Slamming the bytes into a data with a sign is tempting, except someone may try to decode the data as another type, so you have to be able to figure out if the integer fits in the type they asked for. This is a little bit subtle (and we already have all the machinery to do it for strings in the standard library), so it’s tempting to just use string again.